Implement templates for Send Email Rules Engine action
